rename file_viewer.py -> thumbnail_viewer.py
